\name{print.BioMartGOGeneSets_info}
\alias{print.BioMartGOGeneSets_info}
\title{
Print the BioMartGOGeneSets object
}
\description{
Print the BioMartGOGeneSets object
}
\usage{
\method{print}{BioMartGOGeneSets_info}(x, ...)
}
\arguments{

  \item{x}{A \code{BioMartGOGeneSets_info} object.}
  \item{...}{Other arguments}

}
\value{
No value is returned.
}
\examples{
BioMartGOGeneSets
}
