\name{supportedOrganisms}
\alias{supportedOrganisms}
\title{
All supported organisms
}
\description{
All supported organisms
}
\usage{
supportedOrganisms(html = TRUE)
}
\arguments{

  \item{html}{Whether to open the table in the web browser?}

}
\value{
A data frame of supported organisms.
}
\examples{
if(interactive()) {
    supportedOrganisms()
}
}
